new invoice revision complete

+++ app/src/main/java/com/avadroid/avadroid/model/Product.java
@@ -2,11 +2,11 @@ package com.avadroid.avadroid.model;
 
 public class Product {
     public String name;
-    public int price;
+    public float price;
     public int taxRate;
     public int quantity;
 
-    public Product(String name, int price, int taxRate, int quantity) {
+    public Product(String name, float price, int taxRate, int quantity) {
         this.name = name;
         this.price = price;
         this.taxRate = taxRate;
